test Window Case

+++ src/main/scala/com/flink/example/usecase/CaseUtil.scala
@@ -1,5 +1,5 @@
 package com.flink.example.usecase
 
 object CaseUtil {
-  case class GamePlay(gameId : String, uid : String, startTimeStamp : Int, leaveTimeStamp : Int, timeLen : Int, userIp : String)
+  case class GamePlay(gameId : String, uid : String, startTimeStamp : Long, leaveTimeStamp : Long, timeLen : Int, userIp : String)
 }
